backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-jwt-extended==4.5.2
python-dotenv==1.0.0
chromadb==0.4.18
openai==1.3.5
pypdf2==3.0.1
python-magic==0.4.27
boto3==1.28.44
pillow==10.0.0
pytesseract==0.3.10
pandas==2.1.0
matplotlib==3.7.2
numpy==1.24.3
python-multipart==0.0.6
bcrypt==4.0.1
gunicorn==21.2.0